+++ numba/dppl/dppl_lowerer.py
@@ -39,7 +39,7 @@ from numba.core.errors import NumbaParallelSafetyWarning, NumbaPerformanceWarnin
 
 from .target import SPIR_GENERIC_ADDRSPACE
 from .dufunc_inliner import dufunc_inliner
-from . import dppl_host_fn_call_gen1 as dppl_call_gen
+from . import dppl_host_fn_call_gen as dppl_call_gen
 import dpctl
 from numba.dppl.target import DPPLTargetContext
